+++ src/main/java/nl/carsforyou/garage/repositories/ServiceOrderRepository.java
@@ -4,5 +4,5 @@ package nl.carsforyou.garage.repositories;
 import nl.carsforyou.garage.entities.ServiceOrderEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ServiceOrderRepository extends JpaRepository<ServiceOrderEntity, Integer> {
+public interface ServiceOrderRepository extends JpaRepository<ServiceOrderEntity, Long> {
 }
