+++ src/main/java/nl/carsforyou/garage/repositories/VehicleRepository.java
@@ -4,5 +4,5 @@ package nl.carsforyou.garage.repositories;
 import nl.carsforyou.garage.entities.VehicleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VehicleRepository extends JpaRepository<VehicleEntity, Integer> {
+public interface VehicleRepository extends JpaRepository<VehicleEntity, Long> {
 }
